requirements.txt
# Dépendances principales
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
scipy>=1.7.0
PyYAML>=6.0

# Traitement d'image et vidéo
Pillow>=9.0.0
imageio>=2.19.0

# Utilitaires
tqdm>=4.64.0
matplotlib>=3.5.0

# Optionnel pour de meilleures performances
# torch>=1.12.0  # Sera installé automatiquement avec ultralytics
# torchvision>=0.13.0

# Pour les notifications (optionnel)
# requests>=2.28.0
# smtplib  # Inclus dans Python standard

# Pour les logs avancés (optionnel)
# loguru>=0.6.0
